docs: add documentation and warnings for technical helper files
Some users were stumbling into helper scripts, which were meant to
*hide* the technical details from them.

This adds some documentation for the helper scripts (provides the API
at the top of the file), and a large sign to point users to the
correct place to look; i.e. either an .ipynb file or an .Rmd file.