Update mapreduce-pipeline example to add more explanation
Refactored and added documentation to Compute-Pi.ipynb to more
explicitly describe the workflow and how to make a pipeline.
Refactored the map and reduce steps to optionally accept minio
credentials and paths so they can be used both with and without
minio.  Renamed them to average and sample to try to be consistent
with naming within the notebook.